convert validation tests and some other fixes
